+++ src/components/agent/services.ts
@@ -52,7 +52,12 @@ export class AgentServices {
     return payment;
   }
 
-  static async showDeposits(): Promise<Deposit[] | null> {
+  static async showDeposits(depositId?: string): Promise<Deposit[] | null> {
+    if (depositId) {
+      const deposit = await DepositsDAO.getById(depositId);
+      if (!deposit) return null;
+      return [deposit];
+    }
     const deposits = DepositsDAO.index();
     return deposits;
   }