+++ src/db/deposits.ts
@@ -1,4 +1,4 @@
-import { Deposit, PrismaClient } from "@prisma/client";
+import { Deposit, Player, PrismaClient } from "@prisma/client";
 import {
   CreateDepositProps,
   DepositRequest,
@@ -7,6 +7,9 @@ import {
 import { ForbiddenError, NotFoundException } from "@/helpers/error";
 import { hidePassword } from "@/utils/auth";
 import CONFIG from "@/config";
+import { RoledPlayer } from "@/types/response/players";
+import { ERR } from "@/config/errors";
+import { CustomError } from "@/middlewares/errorHandler";
 
 const prisma = new PrismaClient();
 
@@ -14,12 +17,16 @@ export class DepositsDAO {
   /**
    * Create a DB entry for a deposit
    */
-  static create(data: CreateDepositProps): Promise<Deposit> {
+  static async create(
+    data: CreateDepositProps,
+  ): Promise<Deposit & { Player: Player }> {
     try {
-      return prisma.deposit.create({
+      const deposit = await prisma.deposit.create({
         data: { ...data, status: CONFIG.SD.DEPOSIT_STATUS.PENDING },
-        // include: { Player: true },
+        include: { Player: true },
       });
+      deposit.Player = hidePassword(deposit.Player);
+      return deposit;
     } catch (error) {
       throw error;
     } finally {
@@ -52,12 +59,15 @@ export class DepositsDAO {
     }
   }
 
-  static getById(id: string) {
+  static async getById(id: string): Promise<Deposit | null> {
     try {
-      return prisma.deposit.findUnique({
+      const deposit = await prisma.deposit.findUnique({
         where: { id },
         include: { Player: true },
       });
+      if (!deposit) return null;
+      deposit.Player = hidePassword(deposit.Player);
+      return deposit;
     } catch (error) {
       throw error;
     } finally {
@@ -108,7 +118,7 @@ export class DepositsDAO {
         where: {
           status: CONFIG.SD.DEPOSIT_STATUS.VERIFIED,
         },
-        include: { Player: true },
+        include: { Player: { include: { roles: true } } },
       });
     } catch (error) {
       throw error;
@@ -117,9 +127,17 @@ export class DepositsDAO {
     }
   }
 
-  static update(id: string, data: DepositUpdatableProps) {
+  static async update(
+    id: string,
+    data: DepositUpdatableProps,
+  ): Promise<Deposit & { Player: Player }> {
     try {
-      return prisma.deposit.update({ where: { id }, data });
+      const deposit = await prisma.deposit.update({
+        where: { id },
+        data,
+        include: { Player: true },
+      });
+      return deposit;
     } catch (error) {
       throw error;
     } finally {
@@ -138,15 +156,18 @@ export class DepositsDAO {
   }
 
   /**
-   * Ensures deposit exists and belongs to player.
+   * Ensures deposit exists and belongs to authed user or user is agent.
    * @throws if checks fail.
    */
-  static async authorizeTransaction(deposit_id: string, player_id: string) {
+  static async authorizeTransaction(deposit_id: string, player: RoledPlayer) {
     try {
       const deposit = await this.getById(deposit_id);
       if (!deposit) throw new NotFoundException();
 
-      if (deposit.player_id !== player_id)
+      if (
+        deposit.player_id !== player.id &&
+        !player.roles.some((r) => r.name === CONFIG.ROLES.AGENT)
+      )
         throw new ForbiddenError("No autorizado");
 
       return deposit;
@@ -162,13 +183,18 @@ export class DepositsDAO {
    *  - deposit exists and belongs to player
    *  - deposit is not completed or deleted
    *  - deposit is not being confirmed (dirty)
+   *  - another deposit with same tracking_number exists
    *
    * If checks pass, sets dirty flag to true (deposit is being confirmed)
    * @throws if checks fail
    */
-  static async authorizeConfirmation(deposit_id: string, player_id: string) {
+  static async authorizeConfirmation(
+    deposit_id: string,
+    tracking_number: string,
+    player: RoledPlayer,
+  ) {
     try {
-      let deposit = await this.authorizeTransaction(deposit_id, player_id);
+      let deposit = await this.authorizeTransaction(deposit_id, player);
       if (deposit.status === CONFIG.SD.DEPOSIT_STATUS.COMPLETED)
         throw new ForbiddenError(
           "No se pueden modificar depositos completados",
@@ -177,6 +203,14 @@ export class DepositsDAO {
         throw new ForbiddenError("No se pueden modificar depositos eliminados");
       if (deposit.dirty)
         throw new ForbiddenError("El deposito esta siendo confirmado");
+
+      const duplicate = await prisma.deposit.findFirst({
+        where: {
+          tracking_number,
+          NOT: { id: deposit_id },
+        },
+      });
+      if (duplicate) throw new CustomError(ERR.DEPOSIT_ALREADY_EXISTS);
 
       deposit = await prisma.deposit.update({
         where: { id: deposit_id },
@@ -191,7 +225,7 @@ export class DepositsDAO {
     }
   }
 
-  static authorizeDeletion = this.authorizeConfirmation;
+  // static authorizeDeletion = this.authorizeConfirmation;
 
   static async authorizeCreation(request: DepositRequest) {
     try {